move helpers from charm.py to charm_helpers.py
